small cleanups and checks

+++ fenpei/job.py
@@ -245,14 +245,13 @@ class Job(object):
 
 	def cleanup(self, skip_conflicts=False, *args, **kwargs):
 		if self.is_running() or self.is_complete():
-			if not self.queue is None:
-				if not self.queue.force:
-					if skip_conflicts:
-						return False
-					raise AssertionError('you are trying to clean up a job that is running or completed; ' + \
-						'use -f to force this, or -e to skip these jobs (it could also mean that two ' + \
-						'jobs are use the same name and batchname).')
-		if isdir(self.directory):
+			if self.queue is not None and not self.queue.force:
+				if skip_conflicts:
+					return False
+				raise AssertionError('you are trying to clean up a job that is running or completed; ' + \
+					'use -f to force this, or -e to skip these jobs (it could also mean that two ' + \
+					'jobs are use the same name and batchname).')
+		if self.batch_name and isdir(self.directory):
 			rmtree(self.directory, ignore_errors = True)
 			self._log('cleaned up {0:s}'.format(self), level=2)
 			return True
